Securing Server with Shiro
The server stub is extended with an basic shiro authentication part.
Both things are implemented, a way to get acces without credentials and
a the other way with credentials. This is only a stub and roles, passwords
and names are viewable on github. It is for developing only and should be changed
with better encryption and passwords.